requirements.txt
signalwire-agents
python-dotenv
requests
